requirements.txt
asgiref==3.2.3
Django==3.0
image==1.5.27
imutils==0.5.3
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
